Minor changes for better readability.

+++ Project_Dir/src/main/java/com/telerikacademy/carservice/service/CarEventServiceImpl.java
@@ -31,29 +31,36 @@ public class CarEventServiceImpl implements CarEventService {
 
     public List<CarEvent> getAllCarEvents() {
         List<CarEvent> result = carEventRepository.findAll();
-        if (result.size() == 0) throw new DatabaseItemNotFoundException(CAR_EVENT_EXCEPTION);
+        if (result.size() == 0) {
+            throw new DatabaseItemNotFoundException(CAR_EVENT_EXCEPTION);
+        }
         return result;
     }
 
     @Override
     public CarEvent getCarEventByID(long id) {
         CarEvent eventToFind = carEventRepository.findCarEventByCarEventID(id);
-        if (eventToFind == null) throw new DatabaseItemNotFoundException(String.format(CAR_EVENT_BY_ID_EXCEPTION, id));
+        if (eventToFind == null) {
+            throw new DatabaseItemNotFoundException(String.format(CAR_EVENT_BY_ID_EXCEPTION, id));
+        }
         return eventToFind;
     }
 
     @Override
     public List<CarEvent> getCarEventsByCustomerCarID(long id) {
         List<CarEvent> result = carEventRepository.findAllByCustomerCar_CustomerCarID(id);
-        if (result.size() == 0) throw new DatabaseItemNotFoundException(String.format(CAR_EVENT_BY_CAR_ID_EXCEPTION, id));
-
+        if (result.size() == 0) {
+            throw new DatabaseItemNotFoundException(String.format(CAR_EVENT_BY_CAR_ID_EXCEPTION, id));
+        }
         return result;
     }
 
     @Override
     public void addCarEvent(CarEvent carEvent) {
         Set<Procedure> procedures = carEvent.getProcedures();
-        if (procedures.size() == 0) throw new DatabaseItemNotFoundException(PROCEDURE_LIST_EMPTY);
+        if (procedures.size() == 0) {
+            throw new DatabaseItemNotFoundException(PROCEDURE_LIST_EMPTY);
+        }
 
         double totalPrice = calculateCarEventTotalPrice(procedures);
         carEvent.setTotalPrice(totalPrice);
@@ -64,27 +71,28 @@ public class CarEventServiceImpl implements CarEventService {
 
     @Override
     public void deleteCarEvent(long id) {
+        CarEvent eventToDelete = carEventRepository.findCarEventByCarEventID(id);
+        if (eventToDelete == null) {
+            throw new DatabaseItemNotFoundException(String.format(CAR_EVENT_BY_ID_EXCEPTION, id));
+        }
         carEventRepository.deleteById(id);
     }
 
-    //TODO: functionality not tested in frontend. Will probably throw DatabaseItemAlreadyExistsException!!!
     @Override
     public void editCarEvent(CarEvent carEvent, long id) {
-
         CarEvent eventToChange = carEventRepository.findCarEventByCarEventID(id);
-
         if (eventToChange == null) {
             throw new DatabaseItemNotFoundException(String.format(CAR_EVENT_BY_ID_EXCEPTION, id));
         }
 
-
         Set<Procedure> procedures = carEvent.getProcedures();
-        if (procedures.size() == 0) throw new DatabaseItemNotFoundException(PROCEDURE_LIST_EMPTY);
+        if (procedures.size() == 0) {
+            throw new DatabaseItemNotFoundException(PROCEDURE_LIST_EMPTY);
+        }
         double totalPrice = calculateCarEventTotalPrice(procedures);
         eventToChange.setProcedures(carEvent.getProcedures());
         eventToChange.setFinalized(carEvent.getFinalized());
         eventToChange.setTotalPrice(totalPrice);
-
         carEventRepository.save(eventToChange);
     }
 
@@ -96,7 +104,9 @@ public class CarEventServiceImpl implements CarEventService {
 
     private double calculateCarEventTotalPrice(Set<Procedure> procedures) {
         double price = 0.0;
-        if (procedures.size() == 0) throw new DatabaseItemNotFoundException(PROCEDURE_LIST_EMPTY);
+        if (procedures.size() == 0) {
+            throw new DatabaseItemNotFoundException(PROCEDURE_LIST_EMPTY);
+        }
 
         for (Procedure pr : procedures) {
             price += pr.getProcedurePrice();
